+++ lib/utils/styles.ts
@@ -47,4 +47,22 @@ const selectStyles: StylesConfig = {
   }),
 };
 
-export default selectStyles;
+const graphStyles = {
+  layout: {
+    font: {
+      color: "#dee2e6", // bs-body-color
+    },
+    paper_bgcolor: "#121212", // bs-body-bg
+    plot_bgcolor: "#121212", // bs-body-bg
+    xaxis: {
+      gridcolor: "#212529", // bs-gray-900
+      zerolinecolor: "#6c757d", // bs-gray-600
+    },
+    yaxis: {
+      gridcolor: "#212529", // bs-gray-900
+      zerolinecolor: "#6c757d", // bs-gray-600
+    },
+  },
+};
+
+export { selectStyles, graphStyles };
